% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FALCO.R
\name{FALCO}
\alias{FALCO}
\title{Perform factor analysis in data with dependent samples.}
\usage{
FALCO(
  Y,
  K = NULL,
  Z = NULL,
  B = NULL,
  updateDelta = F,
  A.ine = NULL,
  A.equ = NULL,
  c.ine = NULL
)
}
\arguments{
\item{Y}{A \code{p} x \code{n} (number genes/methylation sites x sample size) matrix of observed expression/methylation}

\item{K}{The number of latent factors (an integer >0). If left unspecified, it is estimated using \code{CBCV_plus}.}

\item{Z}{A \code{n} x \code{r} (sample size x number of covariates) matrix of nuisance covariates. This will typically be the intercept (i.e. a vector of all ones). The default value is no covariates.}

\item{B}{An \code{n} x \code{n} positive semidefinite matrix or list of \code{n} x \code{n} positive semidefinite matrices that determine the correlation structure between samples. For example, these can be kinship matrices, partition matrices that group samples by blocks, diagonal matrices with 0's or 1's along the diagonal grouping samples that have the same residual variance, etc. The software includes the identity if it is not in the span of the user-provided matrices. The default is to treat all samples as independent with the same residual variance.}

\item{updateDelta}{\code{T} or \code{F}. If \code{T}, then FALCO performs one iteration PCA assuming Y ~ (LC', Delta, V), where Delta is a diagonal matrix. This tends to improve estimates for C if genomic units have very heterogeneous variances. However, given the computational cost, it is recommended this be set to \code{F} when p >> 10^4 (i.e. DNA methylation data). Defaults to  \code{F}.}

\item{A.ine}{A #inequality constaints x b matrix, where b = length(B) = num. variance components. \code{A.ine \%*\% tau >= 0}, where tau = (tau_1^2,...,tau_b^2) are the variance multipliers. The default is all variance multipliers are greater than 0.}

\item{A.equ}{A #equality constraints x b matrix, where \code{A.equ \%*\% tau = 0}. Defaults to no equality constraints.}

\item{c.ine}{A #inequality constraints vector, where \code{A.ine \%*\% tau - c.ine >= 0}. It is highly recommended users do NOT input anything other than \code{0}. Defaults to \code{0}.}
}
\value{
A list \item{B}{The list of matrices used to parametrize the correlation between samples} \item{v}{The variance multipliers for each B, where v_1B_1 + ... + v_bB_b is the average variance (across genes) between samples} \item{Delta}{If updateDelta=\code{T}, the estimate for Delta in the model Y ~ (LC', Delta, V). If updateDelta=\code{F}, this is \code{NULL}.} \item{Lambda}{Estimates for the non-zero eigenvalues of 1/pL(C'C)L' if updateDelta=\code{F}. If updateDelta=\code{T}, returns estimates for the non-zero eigenvalues of the variance-standardized latent factors, 1/pDelta^\{-1/2\}L(C'C)L'Delta^\{-1/2\}.} \item{C}{The estimate for the latent factors. This has orthogonal columns, and is orthogonal to Z} \item{L}{The estimate for the latent loadings.}
}
\description{
Uses an algorithm called FALCO to estimate the average variance multipliers, latent loadings L, factors C, and eigenvalues of 1/pL(C'C)L'.
}
